Still trying to fix errors. Model only outputs 0% currently

+++ SQL_PHP/database.sql
@@ -25,8 +25,8 @@ CREATE TABLE patientMedicalInfo (
     ever_married ENUM('No', 'Yes') NOT NULL,
     work_type ENUM('children', 'govt_job', 'never_worked', 'private', 'self-employed') NOT NULL,
     residence_type ENUM('Rural', 'Urban') NOT NULL,
-    avg_glucose_level DECIMAL(5,2) CHECK (avg_glucose_level >= 0), 
-    bmi DECIMAL(5,2) CHECK (bmi >= 0), 
+    avg_glucose_level INT CHECK (avg_glucose_level >= 0), 
+    bmi INT CHECK (bmi >= 0), 
     smoking_status ENUM('formerly_smoked', 'never_smoked', 'smokes', 'unknown') NOT NULL,
     stroke VARCHAR(100)
 )
